move to dto
