--- src/components/language/languageName.ts
@@ -1,8 +1,0 @@
-import { type Language } from './dto';
-
-export const languageName = (
-  language: Partial<Pick<Language, 'name' | 'displayName'>>,
-): string | undefined =>
-  language.displayName?.value ||
-  (language.name?.canRead ? language.name.value : undefined) ||
-  undefined;
